Moved SkillColor.java to utils package. Renamed and reorderd the config option to a more fitting place/name.
![screenshotconfigpanel](http://puu.sh/B6AeU/3a8652e146.png)
