+++ runelite-client/src/main/java/net/runelite/client/util/HiscoreLoader.java
